pull code

+++ src/org/nutz/lang/Files.java
@@ -1,7 +1,5 @@
 package org.nutz.lang;
 
-import java.io.BufferedInputStream;
-import java.io.BufferedOutputStream;
 import java.io.BufferedReader;
 import java.io.File;
 import java.io.FileFilter;
@@ -10,8 +8,9 @@ import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
-import java.io.OutputStream;
+import java.io.RandomAccessFile;
 import java.io.Reader;
+import java.nio.channels.FileChannel;
 import java.util.ArrayList;
 import java.util.Enumeration;
 import java.util.LinkedList;
@@ -23,6 +22,7 @@ import java.util.zip.ZipFile;
 import org.nutz.lang.util.Callback;
 import org.nutz.lang.util.ClassTools;
 import org.nutz.lang.util.Disks;
+import org.nutz.log.Logs;
 
 /**
  * 文件操作的帮助函数
@@ -280,11 +280,7 @@ public class Files {
     }
 
     /**
-     * 获取文件后缀名，不包括 '.'，如 'abc.gif','，则返回 'gif'
-     * 
-     * @param f
-     *            文件
-     * @return 文件后缀名
+     * @see #getSuffixName(String)
      */
     public static String getSuffixName(File f) {
         if (null == f)
@@ -307,6 +303,32 @@ public class Files {
         if (-1 == p0 || p0 < p1)
             return "";
         return path.substring(p0 + 1);
+    }
+
+    /**
+     * @see #getSuffix(String)
+     */
+    public static String getSuffix(File f) {
+        if (null == f)
+            return null;
+        return getSuffix(f.getAbsolutePath());
+    }
+
+    /**
+     * 获取文件后缀名，包括 '.'，如 'abc.gif','，则返回 '.gif'
+     * 
+     * @param path
+     *            文件路径
+     * @return 文件后缀
+     */
+    public static String getSuffix(String path) {
+        if (null == path)
+            return null;
+        int p0 = path.lastIndexOf('.');
+        int p1 = path.lastIndexOf('/');
+        if (-1 == p0 || p0 < p1)
+            return "";
+        return path.substring(p0);
     }
 
     /**
@@ -395,10 +417,14 @@ public class Files {
         if (null == thePath)
             thePath = Disks.normalize(path);
         File f = new File(thePath);
-        if (!f.exists())
-            Files.makeDir(f);
+        if (!f.exists()) {
+            boolean flag = Files.makeDir(f);
+            if (!flag) {
+                Logs.get().warnf("create filepool dir(%s) fail!!", f.getPath());
+            }
+        }
         if (!f.isDirectory())
-            throw Lang.makeThrow("'%s' should be a directory!", path);
+            throw Lang.makeThrow("'%s' should be a directory or don't have permission to create it!", path);
         return f;
     }
 
@@ -412,8 +438,11 @@ public class Files {
     public static File createDirIfNoExists(File d) {
         if (null == d)
             return d;
-        if (!d.exists())
-            Files.makeDir(d);
+        if (!d.exists()) {
+            if (!Files.makeDir(d)) {
+                throw Lang.makeThrow("fail to create '%s', permission deny?", d.getAbsolutePath());
+            }
+        }
         if (!d.isDirectory())
             throw Lang.makeThrow("'%s' should be a directory!", d);
         return d;
@@ -472,14 +501,12 @@ public class Files {
         /**
          * 仅文件
          */
-        FILE,
-        /**
-         * 仅目录
-         */
-        DIR,
-        /**
-         * 文件和目录
-         */
+        FILE, /**
+               * 仅目录
+               */
+        DIR, /**
+              * 文件和目录
+              */
         ALL
     }
 
@@ -498,10 +525,7 @@ public class Files {
      * @return 得到文件对象数组
      * @see LsMode
      */
-    public static File[] ls(File d,
-                            final Pattern p,
-                            final boolean exclude,
-                            LsMode mode) {
+    public static File[] ls(File d, final Pattern p, final boolean exclude, LsMode mode) {
         if (null == p) {
             return d.listFiles();
         }
@@ -620,9 +644,7 @@ public class Files {
      * @return 文件对象，如果不存在，则为 null
      */
     public static File findFile(String path) {
-        return findFile(path,
-                        ClassTools.getClassLoader(),
-                        Encoding.defaultEncoding());
+        return findFile(path, ClassTools.getClassLoader(), Encoding.defaultEncoding());
     }
 
     /**
@@ -652,9 +674,7 @@ public class Files {
      * 
      * @return 输出流
      */
-    public static InputStream findFileAsStream(String path,
-                                               Class<?> klass,
-                                               String enc) {
+    public static InputStream findFileAsStream(String path, Class<?> klass, String enc) {
         File f = new File(path);
         if (f.exists())
             try {
@@ -666,9 +686,7 @@ public class Files {
         if (null != klass) {
             InputStream ins = klass.getClassLoader().getResourceAsStream(path);
             if (null == ins)
-                ins = Thread.currentThread()
-                            .getContextClassLoader()
-                            .getResourceAsStream(path);
+                ins = Thread.currentThread().getContextClassLoader().getResourceAsStream(path);
             if (null != ins)
                 return ins;
         }
@@ -775,9 +793,7 @@ public class Files {
         if (null == dir || !dir.exists())
             return false;
         if (!dir.isDirectory())
-            throw new RuntimeException("\""
-                                       + dir.getAbsolutePath()
-                                       + "\" should be a directory!");
+            throw new RuntimeException("\"" + dir.getAbsolutePath() + "\" should be a directory!");
         File[] files = dir.listFiles();
         boolean re = false;
         if (null != files) {
@@ -821,40 +837,86 @@ public class Files {
         if (!dir.exists())
             return false;
         File[] fs = dir.listFiles();
-        for (File f : fs) {
-            if (f.isFile())
-                Files.deleteFile(f);
-            else if (f.isDirectory())
-                Files.deleteDir(f);
+        if (fs != null) {
+            for (File f : fs) {
+                if (f.isFile())
+                    Files.deleteFile(f);
+                else if (f.isDirectory())
+                    Files.deleteDir(f);
+            }
         }
-        return false;
+        return true;
     }
 
     /**
-     * 拷贝一个文件
+     * 相当于 copyFile(src,target,-1)
      * 
-     * @param src
-     *            原始文件
-     * @param target
-     *            新文件
-     * @return 是否拷贝成功
-     * @throws IOException
+     * @see #copyFile(File, File, long)
      */
     public static boolean copyFile(File src, File target) throws IOException {
+        return copyFile(src, target, -1);
+    }
+
+    /**
+     * 将一个文件 copy 一部分（或者全部）到另外一个文件。如果目标文件不存在，创建它先。
+     * 
+     * @param src
+     *            源文件
+     * @param target
+     *            目标文件
+     * @param count
+     *            要 copy 的字节数，0 表示什么都不 copy， -1 表示 copy 全部数据
+     * @return 是否成功
+     * @throws IOException
+     */
+    public static boolean copyFile(File src, File target, long count) throws IOException {
         if (src == null || target == null || !src.exists())
             return false;
         if (!target.exists())
             if (!createNewFile(target))
                 return false;
-        InputStream ins = new BufferedInputStream(new FileInputStream(src));
-        OutputStream ops = new BufferedOutputStream(new FileOutputStream(target));
 
-        Streams.write(ops, ins);
+        // 0 字节？ 那就啥都不做咯
+        if (count == 0)
+            return true;
 
-        Streams.safeClose(ins);
-        Streams.safeFlush(ops);
-        Streams.safeClose(ops);
+        FileInputStream ins = null;
+        FileOutputStream ops = null;
+        FileChannel in = null;
+        FileChannel out = null;
+
+        try {
+            ins = new FileInputStream(src);
+            ops = new FileOutputStream(target);
+            in = ins.getChannel();
+            out = ops.getChannel();
+
+            long maxCount = in.size();
+            if (count < 0 || count > maxCount)
+                count = maxCount;
+
+            in.transferTo(0, count, out);
+        }
+        finally {
+            Streams.safeClose(out);
+            Streams.safeFlush(ops);
+            Streams.safeClose(ops);
+            Streams.safeClose(in);
+            Streams.safeClose(ins);
+        }
         return target.setLastModified(src.lastModified());
+    }
+
+    /**
+     * @see #copyFile(File, File, long)
+     */
+    public static boolean copyFileWithoutException(File src, File target, long count) {
+        try {
+            return copyFile(src, target, -1);
+        }
+        catch (IOException e) {
+            throw Lang.wrapThrow(e);
+        }
     }
 
     /**
@@ -891,8 +953,7 @@ public class Files {
         if (src == null || target == null || !src.exists())
             return false;
         if (!src.isDirectory())
-            throw new IOException(src.getAbsolutePath()
-                                  + " should be a directory!");
+            throw new IOException(src.getAbsolutePath() + " should be a directory!");
         if (!target.exists())
             if (!makeDir(target))
                 return false;
@@ -901,13 +962,9 @@ public class Files {
         if (null != files) {
             for (File f : files) {
                 if (f.isFile())
-                    re &= copyFile(f, new File(target.getAbsolutePath()
-                                               + "/"
-                                               + f.getName()));
+                    re &= copyFile(f, new File(target.getAbsolutePath() + "/" + f.getName()));
                 else
-                    re &= copyDir(f, new File(target.getAbsolutePath()
-                                              + "/"
-                                              + f.getName()));
+                    re &= copyDir(f, new File(target.getAbsolutePath() + "/" + f.getName()));
             }
         }
         return re;
@@ -1020,9 +1077,10 @@ public class Files {
      *            要清除的目录名
      * @throws IOException
      */
-    public static void cleanAllFolderInSubFolderes(File dir, String name)
-            throws IOException {
+    public static void cleanAllFolderInSubFolderes(File dir, String name) throws IOException {
         File[] files = dir.listFiles();
+        if (files == null)
+        	return;
         for (File d : files) {
             if (d.isDirectory())
                 if (d.getName().equalsIgnoreCase(name))
@@ -1039,7 +1097,8 @@ public class Files {
      *            文件1
      * @param f2
      *            文件2
-     * @return <ul>
+     * @return
+     *         <ul>
      *         <li>true: 两个文件内容完全相等
      *         <li>false: 任何一个文件对象为 null，不存在 或内容不相等
      *         </ul>
@@ -1078,7 +1137,7 @@ public class Files {
                 return new File(dir.getAbsolutePath() + "/" + path);
             return new File(dir.getParent() + "/" + path);
         }
-        return new File(path);
+        throw Lang.makeThrow("dir noexists: %s", dir);
     }
 
     /**
@@ -1091,9 +1150,7 @@ public class Files {
     public static File[] dirs(File dir) {
         return dir.listFiles(new FileFilter() {
             public boolean accept(File f) {
-                return !f.isHidden()
-                       && f.isDirectory()
-                       && !f.getName().startsWith(".");
+                return !f.isHidden() && f.isDirectory() && !f.getName().startsWith(".");
             }
         });
     }
@@ -1117,9 +1174,7 @@ public class Files {
     private static void scanDirs(File rootDir, List<File> list) {
         File[] dirs = rootDir.listFiles(new FileFilter() {
             public boolean accept(File f) {
-                return !f.isHidden()
-                       && f.isDirectory()
-                       && !f.getName().startsWith(".");
+                return !f.isHidden() && f.isDirectory() && !f.getName().startsWith(".");
             }
         });
         if (dirs != null) {
@@ -1156,7 +1211,8 @@ public class Files {
      *            文件对象
      * @param f2
      *            文件对象
-     * @return <ul>
+     * @return
+     *         <ul>
      *         <li>true: 两个文件内容完全相等
      *         <li>false: 任何一个文件对象为 null，不存在 或内容不相等
      *         </ul>
@@ -1221,6 +1277,43 @@ public class Files {
         }
         finally {
             Streams.safeClose(br);
+        }
+    }
+    
+    public static int readRange(File f, int pos, byte[] buf, int at, int len) {
+        try {
+            if (f == null || !f.exists())
+                return 0;
+            long fsize = f.length();
+            if (pos > fsize)
+                return 0;
+            len = Math.min(len, buf.length - at);
+            if (pos + len > fsize) {
+                len = (int)(fsize - pos);
+            }
+            RandomAccessFile raf = new RandomAccessFile(f, "r");
+            raf.seek(pos);
+            raf.readFully(buf, at, len);
+            raf.close();
+            return len;
+        }
+        catch (IOException e) {
+            return -1;
+        }
+    }
+    
+    public static int writeRange(File f, int pos, byte[] buf, int at, int len) {
+        try {
+            if (f == null || !f.exists())
+                return 0;
+            RandomAccessFile raf = new RandomAccessFile(f, "rw");
+            raf.seek(pos);
+            raf.write(buf, at, len);
+            raf.close();
+            return len;
+        }
+        catch (IOException e) {
+            return -1;
         }
     }
 }
